% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testWrappers.R
\name{corr}
\alias{corr}
\title{CORR}
\usage{
corr(
  expression,
  annotation = NULL,
  aspects = c("Molecular Function", "Cellular Component", "Biological Process"),
  iterations = 10000,
  geneReplicates = c("mean", "best"),
  pAdjust = c("FDR", "Bonferroni"),
  geneSetDescription = "Latest_GO",
  customGeneSets = NULL,
  minClassSize = 20,
  maxClassSize = 200,
  output = NULL,
  return = TRUE
)
}
\arguments{
\item{expression}{A file path or a data frame. Expression data. (test = CORR only)
Necesary correlation anaylsis. See http://erminej.msl.ubc.ca/help/input-files/gene-expression-profiles/
for data format}

\item{annotation}{Annotation. A file path, a data.frame or a platform short 
name (eg. GPL127). If given a platform short name it will be downloaded
from annotation repository of Pavlidis Lab (\url{https://gemma.msl.ubc.ca/annots/}). 
To get a list of available annotations, use \code{\link{listGemmaAnnotations}}.
Note that if there is a file or folder with the same name as the platform 
name in the directory, that file will be read instead of getting a copy from 
Pavlidis Lab. If this file isn't a valid annotation file, the function will fail.
If providing a custom annotation file, see \code{\link{makeAnnotation}} to do it from
R or  \url{erminej.msl.ubc.ca/help/input-files/gene-annotations/} to do it manually.

If you are providing a custom gene set, you can leave annotation as NULL}

\item{aspects}{Character vector. Which Go aspects to include in the analysis.
Can be in long form (eg. 'Molecular Function') or short form (eg. \code{c('M','C','B')})}

\item{iterations}{Number of iterations. We suggest a starting value of 10000 
iterations. When you decide on parameters you like, we recommend a larger
number of iterations (perhaps 200,000 or more). This is to get sufficient
precision in the p-values to make multiple-test correction work correctly. 
(test = GSR CORR and precRecall methods only)}

\item{geneReplicates}{What to do when genes have multiple scores in input file
(due to multiple probes per gene)}

\item{pAdjust}{Which multiple test correction method to use. Can be "FDR" or
'Westfall-Young' (slower).}

\item{geneSetDescription}{"Latest_GO", a file path that leads to a GO XML or OBO file
or a URL that leads to a go ontology file that ends with rdf-xml.gz. 

If you left annotation as NULL and provided customGeneSets, this argument is
not required and will default to NULL. Otherwise, by default it'll be set to
"Latest_GO" which downloads the latest available GO XML file. This option won't work
without an internet connection. To get a frozen file
that you can use later, see \code{\link{goToday}}, \code{\link{goAtDate}} and \code{\link{getGoDates}}.
 See \url{http://erminej.msl.ubc.ca/help/input-files/gene-set-descriptions/}
for details.}

\item{customGeneSets}{Path to a directory that contains custom gene set files,
 paths to custom gene set files themselves or a named list of character strings.
Use this option to create your own gene sets. If you provide directory you can
specify probes or gene symbols to include in your gene sets. 
See \url{http://erminej.msl.ubc.ca/help/input-files/gene-sets/}
for information about format for this file. If you are providing a list, only gene
symbols are accepted.}

\item{minClassSize}{minimum class size}

\item{maxClassSize}{maximum class size}

\item{output}{Output file name.}

\item{return}{If results should be returned. Set to FALSE if you only want a file}
}
\value{
A list containing a "results" component and a "details" component. 
"results" is a data.frame containing the main output. The columns of this table are

\itemize{
    \item \code{Name}: the name of the gene set
    \item \code{ID}: the id of the gene set
    \item \code{NumProbes}:  the number of elements (e.g. probes) in the gene set.
    \item \code{NumGenes}: the number of genes in the gene set.
    \item \code{RawScore}: the raw statistic for the gene set. For explanations see \href{https://erminej.msl.ubc.ca/help/tutorials/explanation-interpreting-gene-set-scores/}{this} page
    \item \code{Pval}: the p value for the gene set.
    \item \code{CorrectedPvalue}: the corrected p pvalue. See \href{https://erminej.msl.ubc.ca/multiple-test-correction-in-erminej/}{this} page for more information.
    \item \code{MFPvalue}: pvalue after multifunctionality correction. Might be missing if correction was not performed.
    \item \code{CorrectedMFPvalue}: Like CorrectedPvalue, but for the multifunctionality “corrected” pvalue.
    \item \code{Multifunctionality}: How biased the genes in the set are towards \href{https://erminej.msl.ubc.ca/help/tutorials/multifunctionality/}{multifunctional} genes.
    \item \code{Same as}: a list of gene sets which have the exact same members as this one. Such gene sets are not listed anywhere else.
    \item \code{GeneMembers}: If you selected the “Include genes” option when saving, this will contain a list of the genes that are in the gene set, separated by “|”.
}

"details" section contain settings that were used to run the analysis.
}
\description{
This method examines the gene expression profiles themselves,
  not the gene scores for each gene (which is how the other methods like ORA
  work). A score is computed for a gene set based on how correlated the
  expression profiles are.
  
  This can be thought of as a measure of how well the genes in the set
  cluster together, but they need not all be in the same cluster. Thus a gene
  set that contains two coherent clusters that encompass most of the genes in
  the set will tend to get a good score (though not as good as a gene set
  that is just one big cluster).
  
  If you are interested in gene clustering, as opposed to simply looking at
  differntial expression, this method is appropriate. If you feel limited by
  the choice of distance metrics in ermineJ, ORA would be an alternative, but
  you have to define distinct clusters of genes to do that.
  
  One alternative use of correlation scoring is as a control for
  gene-score-based analysis. Correlated gene sets can cause spurious high
  scores, especially if the differential expression in your study is weak. To
  use this approach, you could first use gene-score-based analysis (e.g.,
  gene score resampling and then analyze the data using correlation analysis.
  If any of your gene sets have high scores in both analyses, you should look
  at the data to see if the correlation is not associated with the
  differential expression. This is a simple (but ad hoc) alternative to using
  resampling over the samples to do the gene-score-based analysis.
  
  Method overview taken from: 
  \url{http://erminej.msl.ubc.ca/help/tutorials/running-an-analysis-correlation/}
}
