% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonArguments.R
\name{expression}
\alias{expression}
\title{expression}
\arguments{
\item{expression}{A file path or a data frame. Expression data. (test = CORR only)
Necesary correlation anaylsis. See http://erminej.msl.ubc.ca/help/input-files/gene-expression-profiles/
for data format}
}
\description{
expression
}
\keyword{CORR}
\keyword{internal}
