% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonArguments.R
\name{generalStats}
\alias{generalStats}
\title{generalStats}
\usage{
generalStats(geneReplicates = c("mean", "best"), pAdjust = c("FDR",
  "Westfall-Young"), multifunctionalityCorrection = TRUE)
}
\arguments{
\item{geneReplicates}{What to do when genes have multiple scores in input file
(due to multiple probes per gene)}

\item{pAdjust}{Which multiple test correction method to use. Can be "FDR" or
'Westfall-Young' (slower).}

\item{multifunctionalityCorrection}{Logical. Should the resutls be corrected 
for multifunctionality.}
}
\description{
generalStats
}
\keyword{internal}
\keyword{universal}
