% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscFuns.R
\name{getGemmaAnnot}
\alias{getGemmaAnnot}
\title{Get annotation files from gemma}
\usage{
getGemmaAnnot(chipName, chipFile, annotType = c("noParents",
  "bioProcess", "allParents"), overwrite = FALSE)
}
\arguments{
\item{chipName}{Name of platform. Use \code{\link{listGemmaAnnotations}} to get
a list of valid platform names.}

\item{chipFile}{Name of the output file.}

\item{annotType}{Type of annotation to download. ErmineR typically uses "noParents"}

\item{overwrite}{Should it overwrite an existing file?}
}
\description{
Get annotation files from gemma
}
