% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{getGoDates}
\alias{getGoDates}
\title{getGoDates}
\usage{
getGoDates()
}
\description{
Get all dates for which a version of the GO is available. Any date returned here can be used as an input of
\code{\link{goAtDate}}
}
