% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscFuns.R
\name{getGoGenes}
\alias{getGoGenes}
\title{Get genes for a term}
\usage{
getGoGenes(result, goTerm)
}
\arguments{
\item{result}{An ermineJ output as returned by \code{\link{ermineR}} and it's
wrapper functions (\code{\link{ora}}, \code{\link{gsr}}, \code{\link{corr}},
\code{\link{roc}})}

\item{goTerms}{A go term (eg. 'GO:0007218') or term name ('neuropeptide signaling pathway')}
}
\value{
A vector of gene names, annotated with the GO term
}
\description{
Gets genes for a go term that was available in the analysis
}
