% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{goAtDate}
\alias{goAtDate}
\title{goAtDate}
\usage{
goAtDate(path, date, overwrite = FALSE)
}
\arguments{
\item{path}{File path. Without gz extension}

\item{date}{Character. A valid date as returned from \code{\link{getGoDates}}.}

\item{overwrite}{If TRUE, overwrites existing file}
}
\description{
Get Go terms from a particular date. Use \code{\link{getGoDates}} to see which
dates are available
}
