% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{goToday}
\alias{goToday}
\title{goToday}
\usage{
goToday(path, overwrite = FALSE)
}
\arguments{
\item{path}{File path. Without gz extension}

\item{overwrite}{If TRUE, overwrites existing file}
}
\description{
Downloads latest go ontology
}
