% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonArguments.R
\name{GSRstats}
\alias{GSRstats}
\title{GSRstats}
\arguments{
\item{stats}{Method for computing raw class statistics (test = GSR only)}

\item{quantile}{Integer. Quantile to use. (stats = meanAboveQuantile only)}
}
\description{
GSRstats
}
\keyword{GSR}
\keyword{internal}
