% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonArguments.R
\name{hitlis}
\alias{hitlis}
\alias{hitlist}
\title{hitlist}
\usage{
hitlist(hitlist = NULL)
}
\arguments{
\item{hitlist}{A vector of gene identifiers. ORA method accepts hitlists
instead of scores. If a hitlist is provided, logTrans, thresholds and 
bigIsBetter options are ignored.}
}
\description{
hitlist
}
\keyword{ORA}
\keyword{internal}
