% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonArguments.R
\name{iterations}
\alias{iterations}
\title{iterations}
\usage{
iterations(iterations = 10000)
}
\arguments{
\item{iterations}{Number of iterations. We suggest a starting value of 10000 
iterations. When you decide on parameters you like, we recommend a larger
number of iterations (perhaps 200,000 or more). This is to get sufficient
precision in the p-values to make multiple-test correction work correctly. 
(test = GSR CORR and precRecall methods only)}
}
\description{
iterations
}
\keyword{CORR}
\keyword{GSR}
\keyword{internal}
\keyword{precRecall}
