% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscFuns.R
\name{listGemmaAnnotations}
\alias{listGemmaAnnotations}
\title{List gemma annotations}
\usage{
listGemmaAnnotations()
}
\description{
Returns a vector that has names of all available gemma annotation files that
can be automatically downloaded from Gemma for use with ermineR. See \code{annotation}
argument in \code{\link{ermineR}} and it's wrappers ()
}
