% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{makeAnnotation}
\alias{makeAnnotation}
\title{makeAnnotation}
\usage{
makeAnnotation(
  annotationList,
  symbol = names(annotationList),
  name = NULL,
  output = NULL,
  return = TRUE
)
}
\arguments{
\item{annotationList}{A list of go terms for each gene identifier. Gene identifiers
are the names of the list, each element of the list should have character vectors
with go terms in}

\item{symbol}{Option gene symbols. If not provided IDs from annotation list
will be used}

\item{name}{Optional gene name}

\item{output}{Output file name.}

\item{return}{If results should be returned. Set to FALSE if you only want a file}
}
\description{
Make an annotation object
}
