% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readErmineJOutput.R
\name{readErmineJAnnot}
\alias{readErmineJAnnot}
\title{Read ermineJ annotation files files}
\usage{
readErmineJAnnot(annotation)
}
\arguments{
\item{annotation}{File location}
}
\value{
A data.frame
}
\description{
Read ermineJ annotation files files
}
