% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readErmineJOutput.R
\name{readErmineJOutput}
\alias{readErmineJOutput}
\title{Read ermineJ output files}
\usage{
readErmineJOutput(output)
}
\arguments{
\item{output}{File location}
}
\value{
A data.frame
}
\description{
Read ermineJ output files
}
