% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonArguments.R
\name{returnOpts}
\alias{returnOpts}
\alias{returnOptions}
\title{returnOptions}
\usage{
returnOptions(output = NULL, return = TRUE)
}
\arguments{
\item{output}{Output file name.}

\item{return}{If results should be returned. Set to FALSE if you only want a file}
}
\description{
returnOptions
}
\keyword{internal}
\keyword{universal}
