% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonArguments.R
\name{scores}
\alias{scores}
\title{scores}
\usage{
scores(scores, scoreColumn = 1, bigIsBetter = FALSE, logTrans = FALSE)
}
\arguments{
\item{scores}{A data.frame. Rownames have to be gene identifiers, followed by
any number of columns. The column used for scoring is chosen by \code{scoreColumn}.
See \url{http://erminej.msl.ubc.ca/help/input-files/gene-scores/}
for information abot how to specify scores. (for test = ORA, GSR and ROC)}

\item{scoreColumn}{Integer or character. Which column of the \code{scores} data.frame
to use as scores. Defaults to first column of \code{scores}. See
\url{http://erminej.msl.ubc.ca/help/input-files/gene-scores/} for details.
(for test = ORA, GSR and ROC)}

\item{bigIsBetter}{Logical. If TRUE large scores are considered to be higher.
\code{FALSE} by default (as in p values).}

\item{logTrans}{Logical. Should the data be -log10 transformed. Recommended for 
p values. \code{FALSE} by default}
}
\description{
scores
}
\keyword{GSR}
\keyword{ORA}
\keyword{ROC}
\keyword{internal}
