% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonArguments.R
\name{threshold}
\alias{threshold}
\title{threshold}
\usage{
threshold(threshold = 0.001)
}
\arguments{
\item{threshold}{Double. Score threshold (test = ORA only)}
}
\description{
threshold
}
\keyword{ORA}
\keyword{internal}
