% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonArguments.R
\name{annotation}
\alias{annotation}
\title{annotation}
\usage{
annotation(annotation, aspects = c("Molecular Function", "Cellular Component",
  "Biological Process"))
}
\arguments{
\item{annotation}{Annotation. A file path, a data.frame or a platform short 
name (eg. GPL127). If given a platform short name it will be downloaded
from annotation repository of Pavlidis Lab (\url{www.chibi.ubc.ca/microannots/}). 
Note that if there is a file or folder with the same name as the platform 
name in the directory, that file will be read instead of getting a copy from 
Pavlidis Lab. If this file isn't a valid annotation file, the function will fail.
If providing a custom annotation file, see
\url{erminej.msl.ubc.ca/help/input-files/gene-annotations/}}

\item{aspects}{Character vector. Which Go aspects to include in the analysis}
}
\description{
annotation
}
\keyword{internal}
\keyword{universal}
