% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ermineR.R
\name{ermineR}
\alias{ermineR}
\title{Run ermineJ analysis}
\usage{
ermineR(annotation, scores = NULL, scoreColumn = 2, threshold = 0.001,
  expression = NULL, bigIsBetter = FALSE, customGeneSets = NULL,
  filterNonSpecific = TRUE, geneReplicates = c("mean", "best"),
  genesOut = FALSE, logTrans = FALSE, pAdjust = c("FDR",
  "Westfall-Young"), test = c("ORA", "GSR", "CORR", "ROC"),
  iterations = NULL, stats = c("mean", "quantile", "meanAboveQuantile",
  "precisionRecall"), quantile = 50, geneSetDescription = "Latest_GO",
  multifunctionalityCorrection = TRUE, output = NULL, return = TRUE,
  minClassSize = 10, maxClassSize = 100)
}
\arguments{
\item{annotation}{Annotation. A file path, a data.frame or a platform short 
name (eg. GPL127). If given a platform short name it will be downloaded
from annotation repository of Pavlidis Lab (\url{www.chibi.ubc.ca/microannots/}). 
Note that if there is a file or folder with the same name as the platform 
name in the directory, that file will be read instead of getting a copy from 
Pavlidis Lab. If this file isn't a valid annotation file, the function will fail.
If providing a custom annotation file, see
\url{erminej.msl.ubc.ca/help/input-files/gene-annotations/}}

\item{scores}{A file path or a data.frame. See \url{http://erminej.msl.ubc.ca/help/input-files/gene-scores/}
for information about format for this file. (for test = ORA, GSR and ROC)}

\item{scoreColumn}{Integer. Which column of the \code{scores} includes the scores  (for test = ORA, GSR and ROC)}

\item{threshold}{Double. Score threshold (test = ORA only)}

\item{expression}{A file path or a data frame. Expression data. (test = CORR only)
Necesary correlation anaylsis. See http://erminej.msl.ubc.ca/help/input-files/gene-expression-profiles/
for data format}

\item{bigIsBetter}{Logical. If TRUE large scores are considered to be higher.
\code{FALSE} by default (as in p values).}

\item{customGeneSets}{Directory path or a named list of character strings.
Use this option to create your own gene sets. If you provide directory you can
specify probes or gene symbols to include in your gene sets. 
See \url{http://erminej.msl.ubc.ca/help/input-files/gene-sets/}
for information about format for this file. If you are providing a list, only gene
symbols are accepted.}

\item{filterNonSpecific}{Logical. Filter out non-specific probes}

\item{geneReplicates}{What to do when genes have multiple scores in input file
(due to multiple probes per gene)}

\item{genesOut}{Logical.  Should output include gene symbols for all gene sets}

\item{logTrans}{Logical. Should the data be log transformed. Recommended for 
p values. \code{FALSE} by default}

\item{pAdjust}{Which multiple test correction method to use. Can be "FDR" or
'Westfall-Young' (slower).}

\item{test}{Method for computing gene set significance}

\item{iterations}{Number of iterations (test = GSR and CORR methods only)}

\item{stats}{Method for computing raw class statistics (test = GSR only)}

\item{quantile}{Integer. Quantile to use. (stats = meanAboveQuantile only)}

\item{geneSetDescription}{"Latest_GO", a file path that leads to a GO XML file
or a URL that leads to a go ontology file that ends with rdf-xml.gz. Note that
this is a mandatory argument. It defaults to Latest_GO but that option won't
work if you don't have a working internet connection. See \url{http://erminej.msl.ubc.ca/help/input-files/gene-set-descriptions/}
for details}

\item{multifunctionalityCorrection}{Logical. Should the resutls be corrected 
for multifunctionality.}

\item{output}{Output file name.}

\item{return}{If results should be returned. Set to FALSE if you only want a file}

\item{minClassSize}{minimum class size}

\item{maxClassSize}{maximum class size}
}
\description{
Run ermineJ analysis
}
