% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ermineR.R
\name{ermineR}
\alias{ermineR}
\title{Run ermineJ analysis}
\usage{
ermineR(annotation = NULL, aspects = c("Molecular Function",
  "Cellular Component", "Biological Process"), scores = NULL,
  hitlist = NULL, scoreColumn = 1, threshold = 0.001, expression = NULL,
  bigIsBetter = FALSE, customGeneSets = NULL, geneReplicates = c("mean",
  "best"), logTrans = FALSE, pAdjust = c("FDR", "FWE"), test = c("ORA",
  "GSR", "CORR", "ROC"), iterations = NULL, stats = c("mean", "quantile",
  "meanAboveQuantile", "precisionRecall"), quantile = 50,
  geneSetDescription = "Latest_GO", output = NULL, return = TRUE,
  minClassSize = 20, maxClassSize = 200)
}
\arguments{
\item{annotation}{Annotation. A file path, a data.frame or a platform short 
name (eg. GPL127). If given a platform short name it will be downloaded
from annotation repository of Pavlidis Lab (\url{www.chibi.ubc.ca/microannots/}). 
Note that if there is a file or folder with the same name as the platform 
name in the directory, that file will be read instead of getting a copy from 
Pavlidis Lab. If this file isn't a valid annotation file, the function will fail.
If providing a custom annotation file, see
\url{erminej.msl.ubc.ca/help/input-files/gene-annotations/}}

\item{aspects}{Character vector. Which Go aspects to include in the analysis.
Can be in long form (eg. 'Molecular Function') or short form (eg. \code{c('M','C','B')})}

\item{scores}{A data.frame. Rownames have to be gene identifiers (eg. probes,
must be unique), followed by any number of columns. The column used for
scoring is chosen by \code{scoreColumn}. See
\url{http://erminej.msl.ubc.ca/help/input-files/gene-scores/} for
information abot how to specify scores. (for test = ORA, GSR and ROC)}

\item{scoreColumn}{Integer or character. Which column of the \code{scores} data.frame
to use as scores. Defaults to first column of \code{scores}. See
\url{http://erminej.msl.ubc.ca/help/input-files/gene-scores/} for details.
(for test = ORA, GSR and ROC)}

\item{threshold}{Double. Score threshold (test = ORA only)}

\item{expression}{A file path or a data frame. Expression data. (test = CORR only)
Necesary correlation anaylsis. See http://erminej.msl.ubc.ca/help/input-files/gene-expression-profiles/
for data format}

\item{bigIsBetter}{Logical. If TRUE large scores are considered to be higher.
\code{FALSE} by default (as in p values).}

\item{customGeneSets}{Directory path or a named list of character strings.
Use this option to create your own gene sets. If you provide directory you can
specify probes or gene symbols to include in your gene sets. 
See \url{http://erminej.msl.ubc.ca/help/input-files/gene-sets/}
for information about format for this file. If you are providing a list, only gene
symbols are accepted.}

\item{geneReplicates}{What to do when genes have multiple scores in input file
(due to multiple probes per gene)}

\item{logTrans}{Logical. Should the data be -log10 transformed. Recommended for 
p values. \code{FALSE} by default}

\item{pAdjust}{Which multiple test correction method to use. Can be "FDR" or
'Westfall-Young' (slower).}

\item{test}{Method for computing gene set significance}

\item{iterations}{Number of iterations (test = GSR and CORR methods only)}

\item{stats}{Method for computing raw class statistics (test = GSR only)}

\item{quantile}{Integer. Quantile to use. (stats = meanAboveQuantile only)}

\item{geneSetDescription}{"Latest_GO", a file path that leads to a GO XML or OBO file
or a URL that leads to a go ontology file that ends with rdf-xml.gz. Note that
this is a mandatory argument. It defaults to Latest_GO but that option won't
work if you don't have a working internet connection. See \url{http://erminej.msl.ubc.ca/help/input-files/gene-set-descriptions/}
for details}

\item{output}{Output file name.}

\item{return}{If results should be returned. Set to FALSE if you only want a file}

\item{minClassSize}{minimum class size}

\item{maxClassSize}{maximum class size}
}
\value{
A list
}
\description{
The monolith function that wraps the emrineJ cli. It is recommended to use
the wrapper functions (\code{\link{ora}}, \code{\link{gsr}},
\code{\link{corr}}, \code{\link{roc}}).
}
