% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonArguments.R
\name{geneSetOpts}
\alias{geneSetOpts}
\title{geneSetOpts}
\usage{
geneSetOpts(geneSetDescription = "Latest_GO", customGeneSets = NULL,
  minClassSize = 20, maxClassSize = 200)
}
\arguments{
\item{geneSetDescription}{"Latest_GO", a file path that leads to a GO XML file
or a URL that leads to a go ontology file that ends with rdf-xml.gz. Note that
this is a mandatory argument. It defaults to Latest_GO but that option won't
work if you don't have a working internet connection. See \url{http://erminej.msl.ubc.ca/help/input-files/gene-set-descriptions/}
for details}

\item{customGeneSets}{Directory path or a named list of character strings.
Use this option to create your own gene sets. If you provide directory you can
specify probes or gene symbols to include in your gene sets. 
See \url{http://erminej.msl.ubc.ca/help/input-files/gene-sets/}
for information about format for this file. If you are providing a list, only gene
symbols are accepted.}

\item{minClassSize}{minimum class size}

\item{maxClassSize}{maximum class size}
}
\description{
geneSetOpts
}
\keyword{internal}
\keyword{uniersal}
