% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonArguments.R
\name{iterations}
\alias{iterations}
\title{iterations}
\usage{
iterations(iterations)
}
\arguments{
\item{iterations}{Number of iterations (test = GSR and CORR methods only)}
}
\description{
iterations
}
\keyword{CORR}
\keyword{GSR}
\keyword{internal}
